Fix the types of minlatency and maxlatency...
...so that they can be passed as decimal numbers via Xresources.

+++ config.def.h
@@ -78,8 +78,8 @@ int allowwindowops = 0;
  * near minlatency, but it waits longer for slow updates to avoid partial draw.
  * low minlatency will tear/flicker more, as it can "detect" idle too early.
  */
-static double minlatency = 2;
-static double maxlatency = 33;
+static float minlatency = 2;
+static float maxlatency = 33;
 
 /*
  * Synchronized-Update timeout in ms
@@ -655,8 +655,8 @@ ResourcePref resources[] = {
 		{ "alphaUnfocused",      FLOAT,   &alphaUnfocused },
 		{ "termname",            STRING,  &termname },
 		{ "shell",               STRING,  &shell },
-		{ "minlatency",          INTEGER, &minlatency },
-		{ "maxlatency",          INTEGER, &maxlatency },
+		{ "minlatency",          FLOAT,   &minlatency },
+		{ "maxlatency",          FLOAT,   &maxlatency },
 		{ "su_timeout",          INTEGER, &su_timeout },
 		{ "blinktimeout",        INTEGER, &blinktimeout },
 		{ "bellvolume",          INTEGER, &bellvolume },
